More stone refactoring
