+++ TFC_Shared/src/TFC/API/BlockTypes.java
@@ -1,0 +1,10 @@
+package TFC.API;
+
+public class BlockTypes
+{
+	/* Stone Types */
+	public static final String[] STONE_IGEX = {"Rhyolite", "Basalt", "Andesite", "Dacite"};
+	public static final String[] STONE_IGIN = {"Granite", "Diorite", "Gabbro"};
+	public static final String[] STONE_MM   = {"Quartzite", "Slate", "Phyllite", "Schist", "Gneiss", "Marble"};
+	public static final String[] STONE_SED  = {"Siltstone", "Mudstone", "Shale", "Claystone", "Rock Salt", "Limestone", "Conglomerate", "Dolomite", "Chert", "Chalk"};
+}
