+++ TFC_Shared/src/TFC/Blocks/Terrain/BlockIgEx.java
@@ -15,6 +15,7 @@ import net.minecraft.world.Explosion;
 import net.minecraft.world.World;
 import TFC.TFCBlocks;
 import TFC.TFCItems;
+import TFC.API.BlockTypes;
 import TFC.Core.Helper;
 import TFC.Core.TFC_Core;
 import TFC.Items.Tools.ItemChisel;
@@ -28,7 +29,7 @@ public class BlockIgEx extends BlockStone
 		super(i, material, id);
 		this.dropBlock = TFCBlocks.StoneIgExCobble.blockID;
 
-        names = new String[] {"Rhyolite", "Basalt", "Andesite", "Dacite"};
+        names = BlockTypes.STONE_IGEX;
         icons = new Icon[names.length];
 	}
 
@@ -48,7 +49,6 @@ public class BlockIgEx extends BlockStone
 		if(!world.isRemote)
 		{
 			Random random = new Random();
-
 			ItemStack is = null;
 
 			is = TFC_Core.RandomGem(random, 0);
@@ -59,33 +59,5 @@ public class BlockIgEx extends BlockStone
 				world.spawnEntityInWorld(item);
 			}
 		}
-	}
-
-	/**
-	 * Called when the block is clicked by a player. Args: x, y, z, entityPlayer
-	 */
-	@Override
-	public boolean onBlockActivated(World world, int x, int y, int z, EntityPlayer entityplayer, int par6, float par7, float par8, float par9) 
-	{
-		boolean hasHammer = false;
-		for(int i = 0; i < 9;i++)
-		{
-			if(entityplayer.inventory.mainInventory[i] != null && entityplayer.inventory.mainInventory[i].getItem() instanceof ItemHammer)
-				hasHammer = true;
-		}
-		if(entityplayer.getCurrentEquippedItem() != null && entityplayer.getCurrentEquippedItem().getItem() instanceof ItemChisel && hasHammer && !world.isRemote)
-		{
-			MovingObjectPosition objectMouseOver = Helper.getMouseOverObject(entityplayer, world);
-			if(objectMouseOver == null) {
-				return false;
-			}       
-			int side = objectMouseOver.sideHit;
-
-			int id = world.getBlockId(x, y, z);
-			byte meta = (byte) world.getBlockMetadata(x, y, z);
-
-			return ItemChisel.handleActivation(world, entityplayer, x, y, z, id, meta, side, par7, par8, par9);
-		}
-		return false;
 	}
 }
