+++ TFC_Shared/src/TFC/Blocks/Terrain/BlockMM.java
@@ -15,6 +15,7 @@ import net.minecraft.world.Explosion;
 import net.minecraft.world.World;
 import TFC.TFCBlocks;
 import TFC.TFCItems;
+import TFC.API.BlockTypes;
 import TFC.Core.TFC_Core;
 import TFC.Items.Tools.ItemChisel;
 import TFC.Items.Tools.ItemHammer;
@@ -29,7 +30,7 @@ public class BlockMM extends BlockStone
         super(i, material, id);
         this.dropBlock = TFCBlocks.StoneMMCobble.blockID;
 
-        names = new String[] {"Quartzite", "Slate", "Phyllite", "Schist", "Gneiss", "Marble"};
+        names = BlockTypes.STONE_MM;
         icons = new Icon[names.length];
     }
 
@@ -71,25 +72,4 @@ public class BlockMM extends BlockStone
 
         //mc.ingameGUI.addChatMessage("Meta="+(new StringBuilder()).append(getBlockName()).append(":").append(metadata).toString());  
     }
-    
-    @Override
-    public boolean onBlockActivated(World world, int x, int y, int z, EntityPlayer entityplayer, int side, float par7, float par8, float par9) 
-    {
-        boolean hasHammer = false;
-        for(int i = 0; i < 9;i++)
-        {
-            if(entityplayer.inventory.mainInventory[i] != null && entityplayer.inventory.mainInventory[i].getItem() instanceof ItemHammer)
-                hasHammer = true;
-        }
-        if(entityplayer.getCurrentEquippedItem() != null && entityplayer.getCurrentEquippedItem().getItem() instanceof ItemChisel && hasHammer && !world.isRemote)
-        {
-            int id = world.getBlockId(x, y, z);
-            byte meta = (byte) world.getBlockMetadata(x, y, z);
-
-            return ItemChisel.handleActivation(world, entityplayer, x, y, z, id, meta, side, par7, par8, par9);
-        }
-        return false;
-    }
-
 }
-
