+++ TFC_Shared/src/TFC/Blocks/Terrain/BlockMMBrick.java
@@ -53,7 +53,7 @@ public class BlockMMBrick extends BlockMMSmooth
 	@Override
 	public void registerIcons(IconRegister iconRegisterer)
     {
-		for(int i = 0; i < 6; i++)
+		for(int i = 0; i < names.length; i++)
 		{
 			icons[i] = iconRegisterer.registerIcon("rocks/"+names[i]+" Brick");
 		}
