+++ TFC_Shared/src/TFC/Blocks/Terrain/BlockMMSmooth.java
@@ -3,6 +3,7 @@ package TFC.Blocks.Terrain;
 import java.util.List;
 import java.util.Random;
 import TFC.TerraFirmaCraft;
+import TFC.API.BlockTypes;
 import TFC.Blocks.BlockTerra;
 import TFC.Core.Helper;
 import TFC.Core.TFC_Core;
@@ -44,82 +45,14 @@ import net.minecraft.world.biome.*;
 import net.minecraft.world.chunk.*;
 import net.minecraft.world.gen.feature.*;
 
-public class BlockMMSmooth extends BlockTerra
+public class BlockMMSmooth extends BlockSmooth
 {
 	public BlockMMSmooth(int i) 
 	{
 		super(i, Material.rock);
 		this.setCreativeTab(CreativeTabs.tabBlock);
+
+        names = BlockTypes.STONE_MM;
+		icons = new Icon[names.length];
 	}
-
-	@SideOnly(Side.CLIENT)
-    @Override
-    /**
-     * returns a list of blocks with the same ID, but different meta (eg: wood returns 4 blocks)
-     */
-    public void getSubBlocks(int par1, CreativeTabs par2CreativeTabs, List list)
-    {
-        for(int i = 0; i < 6; i++) {
-            list.add(new ItemStack(this,1,i));
-        }
-    }
-
-    /*
-     * Mapping from metadata value to damage value
-     */
-    @Override
-    public int damageDropped(int i) 
-    {
-        return i;
-    }
-
-    @Override
-	public Icon getIcon(int i, int j) 
-	{
-		return icons[j];
-	}
-
-	protected Icon[] icons = new Icon[6];
-	protected String[] names = {"Quartzite", "Slate", "Phyllite", "Schist", "Gneiss", "Marble"};
-	
-	@Override
-	public void registerIcons(IconRegister iconRegisterer)
-    {
-		for(int i = 0; i < 6; i++)
-		{
-			icons[i] = iconRegisterer.registerIcon("rocks/"+names[i]+" Smooth");
-		}
-    }
-	
-	/**
-     * Called when the block is clicked by a player. Args: x, y, z, entityPlayer
-     */
-    /**
-     * Called when the block is clicked by a player. Args: x, y, z, entityPlayer
-     */
-    @Override
-    public boolean onBlockActivated(World world, int x, int y, int z, EntityPlayer entityplayer, int par6, float par7, float par8, float par9) 
-    {
-        boolean hasHammer = false;
-        for(int i = 0; i < 9;i++)
-        {
-            if(entityplayer.inventory.mainInventory[i] != null && entityplayer.inventory.mainInventory[i].getItem() instanceof ItemHammer)
-                hasHammer = true;
-        }
-        if(entityplayer.getCurrentEquippedItem() != null && entityplayer.getCurrentEquippedItem().getItem() instanceof ItemChisel && hasHammer && !world.isRemote)
-        {
-            MovingObjectPosition objectMouseOver = Helper.getMouseOverObject(entityplayer, world);
-            if(objectMouseOver == null) {
-                return false;
-            }       
-            int side = objectMouseOver.sideHit;
-
-            int id = world.getBlockId(x, y, z);
-            byte meta = (byte) world.getBlockMetadata(x, y, z);
-
-            return ItemChisel.handleActivation(world, entityplayer, x, y, z, id, meta, side, par7, par8, par9);
-        }
-        return false;
-    }
 }
-
