+++ TFC_Shared/src/TFC/Blocks/Terrain/BlockSed.java
@@ -15,6 +15,7 @@ import net.minecraft.world.Explosion;
 import net.minecraft.world.World;
 import TFC.TFCBlocks;
 import TFC.TFCItems;
+import TFC.API.BlockTypes;
 import TFC.Core.Helper;
 import TFC.Core.TFC_Core;
 import TFC.Items.Tools.ItemChisel;
@@ -29,7 +30,7 @@ public class BlockSed extends BlockStone
         super(i, material, id);
         this.dropBlock = TFCBlocks.StoneSedCobble.blockID;
 
-        names = new String[] {"Siltstone", "Mudstone", "Shale", "Claystone", "Rock Salt", "Limestone", "Conglomerate", "Dolomite", "Chert", "Chalk"};
+        names = BlockTypes.STONE_SED;
         icons = new Icon[names.length];
     }
 
@@ -49,7 +50,6 @@ public class BlockSed extends BlockStone
         if(!world.isRemote)
         {
             Random random = new Random();
-
             ItemStack is = null;
 
             is = TFC_Core.RandomGem(random,1);
@@ -62,33 +62,4 @@ public class BlockSed extends BlockStone
 
         }
     }
-    
-    /**
-     * Called when the block is clicked by a player. Args: x, y, z, entityPlayer
-     */
-    @Override
-    public boolean onBlockActivated(World world, int x, int y, int z, EntityPlayer entityplayer, int par6, float par7, float par8, float par9) 
-    {
-        boolean hasHammer = false;
-        for(int i = 0; i < 9;i++)
-        {
-            if(entityplayer.inventory.mainInventory[i] != null && entityplayer.inventory.mainInventory[i].getItem() instanceof ItemHammer)
-                hasHammer = true;
-        }
-        if(entityplayer.getCurrentEquippedItem() != null && entityplayer.getCurrentEquippedItem().getItem() instanceof ItemChisel && hasHammer && !world.isRemote)
-        {
-            MovingObjectPosition objectMouseOver = Helper.getMouseOverObject(entityplayer, world);
-            if(objectMouseOver == null) {
-                return false;
-            }       
-            int side = objectMouseOver.sideHit;
-
-            int id = world.getBlockId(x, y, z);
-            byte meta = (byte) world.getBlockMetadata(x, y, z);
-
-            return ItemChisel.handleActivation(world, entityplayer, x, y, z, id, meta, side, par7, par8, par9);
-        }
-        return false;
-    }
 }
-
