+++ TFC_Shared/src/TFC/Blocks/Terrain/BlockSmooth.java
@@ -1,0 +1,93 @@
+package TFC.Blocks.Terrain;
+
+import java.util.List;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
+import net.minecraft.block.material.Material;
+import net.minecraft.client.renderer.texture.IconRegister;
+import net.minecraft.creativetab.CreativeTabs;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.Icon;
+import net.minecraft.util.MovingObjectPosition;
+import net.minecraft.world.World;
+import TFC.Blocks.BlockTerra;
+import TFC.Core.Helper;
+import TFC.Items.Tools.ItemChisel;
+import TFC.Items.Tools.ItemHammer;
+
+public class BlockSmooth extends BlockTerra
+{
+
+	protected BlockSmooth(int i, Material material)
+	{
+		super(i, material);
+	}
+
+	protected String[] names;
+	protected Icon[] icons;
+	
+	@SideOnly(Side.CLIENT)
+    @Override
+    /**
+     * returns a list of blocks with the same ID, but different meta (eg: wood returns 4 blocks)
+     */
+    public void getSubBlocks(int par1, CreativeTabs par2CreativeTabs, List list)
+	{
+		for(int i = 0; i < names.length; i++) {
+			list.add(new ItemStack(this,1,i));
+		}
+	}
+
+	/*
+	 * Mapping from metadata value to damage value
+	 */
+	@Override
+	public int damageDropped(int i) {
+		return i;
+	}
+
+	@Override
+	public Icon getIcon(int i, int j) 
+	{
+		return icons[j];
+	}
+
+	@Override
+	public void registerIcons(IconRegister iconRegisterer)
+    {
+		for(int i = 0; i < names.length; i++)
+		{
+			icons[i] = iconRegisterer.registerIcon("rocks/"+names[i]+" Smooth");
+		}
+    }
+
+    /**
+     * Called when the block is clicked by a player. Args: x, y, z, entityPlayer
+     */
+    @Override
+    public boolean onBlockActivated(World world, int x, int y, int z, EntityPlayer entityplayer, int par6, float par7, float par8, float par9)  
+    {
+        boolean hasHammer = false;
+        for(int i = 0; i < 9;i++)
+        {
+            if(entityplayer.inventory.mainInventory[i] != null && entityplayer.inventory.mainInventory[i].getItem() instanceof ItemHammer)
+                hasHammer = true;
+        }
+        if(entityplayer.getCurrentEquippedItem() != null && entityplayer.getCurrentEquippedItem().getItem() instanceof ItemChisel && hasHammer && !world.isRemote)
+        {
+            MovingObjectPosition objectMouseOver = Helper.getMouseOverObject(entityplayer, world);
+            if(objectMouseOver == null) {
+                return false;
+            }       
+            int side = objectMouseOver.sideHit;
+
+            int id = world.getBlockId(x, y, z);
+            byte meta = (byte) world.getBlockMetadata(x, y, z);
+
+            return ItemChisel.handleActivation(world, entityplayer, x, y, z, id, meta, side, par7, par8, par9);
+        }
+        return false;
+    }
+}
